Move Tabs and delete Stats
