--- frontend/src/components/stats/index.tsx
@@ -1,3 +1,0 @@
-import { Stats } from "./Stats";
-
-export { Stats };
