+++ frontend/src/views/groups/GroupsView.tsx
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 import { Table } from "../../components/table";
-import { Tabs } from "../../components/tabs";
+import { Tabs } from "./tabs";
 import { GROUPS_URL, getGroupLeaderboardUrl } from "../../helpers/api";
 import axios, { AxiosResponse } from "axios";
 import { Group } from "../../helpers/types";
