+++ frontend/src/views/groups/tabs/Tabs.tsx
@@ -1,8 +1,8 @@
 import { PlusIcon } from "@heroicons/react/24/outline";
-import { Modal } from "../modal/Modal";
+import { Modal } from "../../../components/modal/Modal";
 import { Fragment, useEffect, useRef, useState } from "react";
 import { Transition } from "@headlessui/react";
-import { Group } from "../../helpers/types";
+import { Group } from "../../../helpers/types";
 import { TabItem } from "./TabItem";
 import { SkeletonTabItem } from "./SkeletonTabItem";
 import {
