--- frontend/src/views/hero/AuthenticatedHero.tsx
@@ -1,7 +1,0 @@
-import { Stats } from "../../components/stats";
-
-export const AuthenticatedHero = () => (
-  <section className="h-full min-h-[420px] w-full bg-[url('/online.png')] bg-cover bg-center bg-no-repeat">
-    <Stats />
-  </section>
-);
