+++ frontend/src/views/hero/index.tsx
@@ -1,4 +1,3 @@
-import { AuthenticatedHero } from "./AuthenticatedHero";
 import { DefaultHero } from "./DefaultHero";
 
-export { AuthenticatedHero, DefaultHero };
+export { DefaultHero };
